[front/user/join] wip : tweaked store part by adding persist feature on original store
