+++ front/pages/_app.tsx
@@ -1,14 +1,15 @@
-import '../styles/globals.css'
-import type { AppProps } from 'next/app'
-import { Provider } from 'react-redux'
-import store from '../src/modules/store'
+import '../styles/globals.css';
+import type { AppProps } from 'next/app';
+import { Provider } from 'react-redux';
+// import store from '../src/modules/store';
+import Store from '../src/modules/store';
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
-    <Provider store={store}>
+    <Store>
       <Component {...pageProps} />
-    </Provider>
-  )
+    </Store>
+  );
 }
 
-export default MyApp
+export default MyApp;
